requirements.txt
blinker==1.9.0
click==8.1.7
colorama==0.4.6
et_xmlfile==2.0.0
Flask==2.3.3
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.3
openpyxl==3.1.2
pandas==2.2.3
pyodbc==5.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
SQLAlchemy==2.0.35
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.1.3
